api/pyproject.toml
[project]
name = "gitscapeapi"
version = "0.1.0"
description = "The official API for GitScape, a tool for generating digests from any git repositories."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn>=0.34.2",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "slowapi>=0.1.9",
]
